add mobile stylings to explore tables
