+++ src/components/Global/Explore/DexTokens/DexTokens.module.css
@@ -17,6 +17,8 @@
     height: calc(100vh - 64px);
     overflow-y: hidden;
     background: var(--dark1);
+    border: 1px solid var(--dark3);
+    border-radius: 4px;
    
 }
 
@@ -104,10 +106,7 @@
 
 
  .poolName{
-   /* justify-content: flex-start !important; */
-   width: 30px;
-  
-  
+   justify-content: flex-start !important; 
 }
 .no_results{
     width: 100%;
@@ -150,6 +149,10 @@
 
 
 @media only screen and (min-width: 768px) {
+
+    .mainContainer{
+        border: none;
+    }
     .no_results{
         height: 70%;
     }
@@ -168,7 +171,7 @@
 }
 
 
-.tradeButton, .poolName{
+.tradeButton{
     width: 70px;
 }
 }
