+++ src/components/Global/Explore/DexTokens/DexTokens.tsx
@@ -1,27 +1,24 @@
 import { Dispatch, memo, SetStateAction, useContext } from 'react';
-import Spinner from '../Spinner/Spinner';
-import {
-    ShadowBox,
-    SpinnerContainer,
-    Table,
-    TableBody,
-} from '../../../styled/Components/Analytics';
+import Spinner from '../../Spinner/Spinner';
+import { SpinnerContainer } from '../../../../styled/Components/Analytics';
 import styles from './DexTokens.module.css';
-import { FlexContainer } from '../../../styled/Common';
-import TokenRow from './TokenRow';
-import { useSortedDexTokens, sortedDexTokensIF } from './useSortedDexTokens';
-import { dexTokenData } from '../../../pages/Explore/useTokenStats';
-import TableHeadTokens from './TableHeadTokens';
-import { getDefaultPairForChain } from '../../../ambient-utils/constants';
-import { GCServerPoolIF, PoolIF, TokenIF } from '../../../ambient-utils/types';
-import { PoolContext } from '../../../contexts/PoolContext';
-import useMediaQuery from '../../../utils/hooks/useMediaQuery';
-import { usePoolList2 } from '../../../App/hooks/usePoolList2';
-import { CrocEnvContext } from '../../../contexts/CrocEnvContext';
-import { isWrappedNativeToken } from '../../../ambient-utils/dataLayer';
-import { sortType } from './useSortedPools';
-import AssignSort from './AssignSort';
-import TooltipComponent from '../TooltipComponent/TooltipComponent';
+import { useSortedDexTokens, sortedDexTokensIF } from '../useSortedDexTokens';
+import { dexTokenData } from '../../../../pages/Explore/useTokenStats';
+import { getDefaultPairForChain } from '../../../../ambient-utils/constants';
+import {
+    GCServerPoolIF,
+    PoolIF,
+    TokenIF,
+} from '../../../../ambient-utils/types';
+import { PoolContext } from '../../../../contexts/PoolContext';
+import useMediaQuery from '../../../../utils/hooks/useMediaQuery';
+import { usePoolList2 } from '../../../../App/hooks/usePoolList2';
+import { CrocEnvContext } from '../../../../contexts/CrocEnvContext';
+import { isWrappedNativeToken } from '../../../../ambient-utils/dataLayer';
+import AssignSort from '../AssignSort';
+import TooltipComponent from '../../TooltipComponent/TooltipComponent';
+import TokenRow from '../TokenRow/TokenRow';
+import ExploreToggle from '../ExploreToggle/ExploreToggle';
 
 export type columnSlugs =
     | 'token'
@@ -42,7 +39,7 @@ export interface HeaderItem {
     onClick?: () => void;
     tooltipText?: string | JSX.Element;
     classname?: boolean;
-} 
+}
 
 interface propsIF {
     dexTokens: dexTokenData[];
@@ -50,10 +47,12 @@ interface propsIF {
     goToMarket: (tknA: string, tknB: string) => void;
     searchQuery: string;
     setSearchQuery: Dispatch<SetStateAction<string>>;
+    view: 'pools' | 'tokens';
+    handleToggle(): void
 }
 
 function DexTokens(props: propsIF) {
-    const { dexTokens, chainId, goToMarket, searchQuery, setSearchQuery } =
+    const { dexTokens, chainId, goToMarket, searchQuery, setSearchQuery, view, handleToggle } =
         props;
 
     const { findPool } = useContext(PoolContext);
@@ -76,25 +75,26 @@ function DexTokens(props: propsIF) {
 
     const dexTokensHeaderItems: (HeaderItem | null)[] = [
         // mobileScrenView ? null :
-            {
+        {
             label: 'Token',
             slug: 'token',
             hidden: false,
             align: 'left',
             responsive: 'sm',
             sortable: false,
-            classname: styles.tokens
+            classname: styles.tokens,
         },
-        desktopView ? 
-            {
-            label: 'Name',
-            slug: 'name',
-            hidden: smallScreen,
-            align: 'left',
-            responsive: 'sm',
-                sortable: true,
-            classname: styles.poolName
-        }: null,
+        desktopView
+            ? {
+                  label: 'Name',
+                  slug: 'name',
+                  hidden: smallScreen,
+                  align: 'left',
+                  responsive: 'sm',
+                  sortable: true,
+                  classname: styles.poolName,
+              }
+            : null,
         {
             label: 'Volume',
             slug: 'volume',
@@ -135,22 +135,22 @@ function DexTokens(props: propsIF) {
     const headerDisplay = (
         <div className={styles.headerContainer}>
             {dexTokensHeaderItems
-                .filter((item) => item !== null)
+                .filter((item): item is HeaderItem => item !== null)
                 .map((item: HeaderItem) => {
                     const isActiveSort: boolean =
                         sortedTokens.sortBy.slug === item.slug;
                     return (
                         <div
-                            key={JSON.stringify(item?.label)}
-                            className={`${styles.gridHeaderItem} ${item?.classname} ${styles.headerItems}`}
+                            key={JSON.stringify(item.label)} // No need for optional chaining
+                            className={`${styles.gridHeaderItem} ${item.classname} ${styles.headerItems}`}
                             style={{
-                                cursor: item?.sortable ? 'pointer' : 'default',
+                                cursor: item.sortable ? 'pointer' : 'default',
                             }}
                             onClick={() =>
                                 item.sortable && sortedTokens.update(item.slug)
                             }
                         >
-                            {item?.label}
+                            {item.label}
                             {isActiveSort && (
                                 <AssignSort
                                     direction={
@@ -160,7 +160,7 @@ function DexTokens(props: propsIF) {
                                     }
                                 />
                             )}
-                            {item?.tooltipText && (
+                            {item.tooltipText && (
                                 <TooltipComponent
                                     title={item.tooltipText}
                                     placement='right'
@@ -181,6 +181,8 @@ function DexTokens(props: propsIF) {
 
     return (
         <div className={styles.mainContainer}>
+                        <ExploreToggle view={view} handleToggle={handleToggle}/>
+
             {headerDisplay}
             <div className={`${styles.contentContainer} custom_scroll_ambient`}>
                 <div className={styles.borderRight} />
@@ -224,7 +226,9 @@ function DexTokens(props: propsIF) {
                             />
                         );
                     })
-                ) : searchQuery ? noResults : (
+                ) : searchQuery ? (
+                    noResults
+                ) : (
                     <SpinnerContainer
                         fullHeight
                         fullWidth
@@ -238,88 +242,5 @@ function DexTokens(props: propsIF) {
         </div>
     );
 }
-
-//     return (
-//         <FlexContainer
-//             fullWidth
-//             height={
-//                 showMobileVersion
-//                     ? 'calc(100svh - 240px)'
-//                     : 'calc(100svh - 200px)'
-//             }
-//         >
-//             <div className='custom_scroll_ambient'>
-//                 <ShadowBox>
-//                     <Table>
-//                         <TableHeadTokens
-//                             headerItems={dexTokensHeaderItems}
-//                             sortedTokens={sortedTokens}
-//                         />
-//                         <TableBody>
-//                         {!sortedTokens.data.length && searchQuery ? noResults :
-//                             sortedTokens.data.length ? (
-//                                 sortedTokens.data.map((token: dexTokenData) => {
-//                                     const samplePool: PoolIF | undefined =
-//                                         findPool(
-//                                             token.tokenAddr,
-//                                             defaultTokensForChain[0],
-//                                         ) ??
-//                                         findPool(
-//                                             token.tokenAddr,
-//                                             defaultTokensForChain[1],
-//                                         ) ??
-//                                         findPool(token.tokenAddr);
-//                                     const backupPool:
-//                                         | GCServerPoolIF
-//                                         | undefined = unfilteredPools.find(
-//                                         (p: GCServerPoolIF) =>
-//                                             (p.base.toLowerCase() ===
-//                                                 token.tokenAddr.toLowerCase() &&
-//                                                 !isWrappedNativeToken(
-//                                                     p.quote,
-//                                                 )) ||
-//                                             (p.quote.toLowerCase() ===
-//                                                 token.tokenAddr.toLowerCase() &&
-//                                                 !isWrappedNativeToken(p.base)),
-//                                     );
-//                                     if (
-//                                         !token.tokenMeta ||
-//                                         (!samplePool && !backupPool)
-//                                     )
-//                                         return null;
-//                                     return (
-//                                         <TokenRow
-//                                             key={token.tokenAddr}
-//                                             token={token}
-//                                             tokenMeta={token.tokenMeta}
-//                                             samplePool={samplePool}
-//                                             backupPool={backupPool}
-//                                             goToMarket={goToMarket}
-//                                             smallScreen={smallScreen}
-//                                         />
-//                                     );
-//                                 })
-//                             ) :
-//                                 (
-//                                 <SpinnerContainer
-//                                     fullHeight
-//                                     fullWidth
-//                                     alignItems='center'
-//                                     justifyContent='center'
-//                                 >
-//                                     <Spinner
-//                                         size={100}
-//                                         bg='var(--dark1)'
-//                                         centered
-//                                     />
-//                                 </SpinnerContainer>
-//                             )}
-//                         </TableBody>
-//                     </Table>
-//                 </ShadowBox>
-//             </div>
-//         </FlexContainer>
-//     );
-// }
 
 export default memo(DexTokens);
