+++ src/components/Global/Explore/ExploreToggle/ExploreToggle.module.css
@@ -1,0 +1,21 @@
+
+
+
+.container button{
+    background: transparent;
+    outline: none;
+    border: none;
+    color: var(--text3);
+    font-size: 13px;
+font-style: normal;
+font-weight: 500;
+border-bottom: 1px solid transparent;
+width: 150px;
+cursor: pointer;
+padding: 4px;
+}
+
+.activeButton{
+   color: var(--accent1) !important ;
+   border-bottom: 1px solid var(--accent1) !important;
+}
