+++ src/components/Global/Explore/PoolRow/PoolRow.module.css
@@ -5,7 +5,7 @@
     flex-direction: row;
     align-items: center;
     min-width: 100%;
-    min-height: 64px;
+    min-height: 50px;
    
     gap: 4px;
