+++ src/components/Global/Explore/TokenRow/TokenRow.module.css
@@ -5,7 +5,7 @@
     flex-direction: row;
     align-items: center;
     min-width: 100%;
-    min-height: 64px;
+    min-height: 40px;
    
     gap: 4px;
