+++ src/components/Global/Explore/useSortedDexTokens.ts
@@ -1,6 +1,6 @@
 import { useMemo, useState } from 'react';
 import { dexTokenData } from '../../../pages/Explore/useTokenStats';
-import { columnSlugs } from './DexTokens';
+import { columnSlugs } from './DexTokens/DexTokens';
 
 interface sortIF {
     slug: columnSlugs | null;
