+++ src/pages/Explore/Explore.tsx
@@ -1,12 +1,8 @@
 import { useContext, useEffect, useRef, useState } from 'react';
-import TopPools from '../../components/Global/Explore/TopPools';
-import DexTokens from '../../components/Global/Explore/DexTokens';
 import { ExploreContext } from '../../contexts/ExploreContext';
 import { CrocEnvContext } from '../../contexts/CrocEnvContext';
 import { PoolContext } from '../../contexts/PoolContext';
 import { ChainDataContext } from '../../contexts/ChainDataContext';
-import Toggle from '../../components/Form/Toggle';
-import { Text } from '../../styled/Common';
 import { linkGenMethodsIF, useLinkGen } from '../../utils/hooks/useLinkGen';
 import { AiOutlineDollarCircle } from 'react-icons/ai';
 import { DefaultTooltip } from '../../components/Global/StyledTooltip/StyledTooltip';
@@ -14,6 +10,8 @@ import { PoolIF } from '../../ambient-utils/types';
 import styles from './Explore.module.css';
 import { LuRefreshCcw, LuSearch } from 'react-icons/lu';
 import useOnClickOutside from '../../utils/hooks/useOnClickOutside';
+import TopPools from '../../components/Global/Explore/TopPools/TopPools';
+import DexTokens from '../../components/Global/Explore/DexTokens/DexTokens';
 
 interface ExploreIF {
     view: 'pools' | 'tokens';
@@ -150,24 +148,24 @@ export default function Explore(props: ExploreIF) {
                   );
               })
             : tokens.data;
-    
-            const searchInputRef = useRef<HTMLDivElement>(null);
 
-            const clickOutsideInputHandler = () => {
-                const searchQuery = view === 'pools' ? searchQueryPool : searchQueryToken;
-                const filteredItems = view === 'pools' ? filteredPools : filteredTokens;
-                const setSearchQuery = view === 'pools' ? setSearchQueryPool : setSearchQueryToken;
-            
-                if (!searchQuery || filteredItems.length) {
-                    return null;
-                }
-            
-                setSearchQuery('');
-            };
-            
-            useOnClickOutside(searchInputRef, clickOutsideInputHandler);
+    const searchInputRef = useRef<HTMLDivElement>(null);
 
-            
+    const clickOutsideInputHandler = () => {
+        const searchQuery =
+            view === 'pools' ? searchQueryPool : searchQueryToken;
+        const filteredItems = view === 'pools' ? filteredPools : filteredTokens;
+        const setSearchQuery =
+            view === 'pools' ? setSearchQueryPool : setSearchQueryToken;
+
+        if (!searchQuery || filteredItems.length) {
+            return null;
+        }
+
+        setSearchQuery('');
+    };
+
+    useOnClickOutside(searchInputRef, clickOutsideInputHandler);
 
     const inputContainer = (
         <div className={styles.input_container} ref={searchInputRef}>
@@ -190,88 +188,80 @@ export default function Explore(props: ExploreIF) {
         </div>
     );
 
+    const optionsContent = (
+        <div className={styles.options_content}>
+        {inputContainer}
+        {view === 'pools' && (
+            <DefaultTooltip
+                interactive
+                title={
+                    isExploreDollarizationEnabled
+                        ? 'Switch to prices in native currency'
+                        : 'Switch to prices in USD'
+                }
+                enterDelay={500}
+            >
+                <div className={styles.refresh_container}>
+                    <button
+                        className={styles.refresh_button}
+                        onClick={() =>
+                            setIsExploreDollarizationEnabled(
+                                (prev) => !prev,
+                            )
+                        }
+                    >
+                        {
+                            <AiOutlineDollarCircle
+                                size={20}
+                                id='trade_dollarized_prices_button'
+                                aria-label='Toggle dollarized prices button'
+                                style={{
+                                    color: isExploreDollarizationEnabled
+                                        ? 'var(--accent1)'
+                                        : undefined,
+                                }}
+                            />
+                        }
+                    </button>
+                </div>
+            </DefaultTooltip>
+        )}
+        <DefaultTooltip
+            interactive
+            title={
+                view === 'pools'
+                    ? 'Refresh Top Pools'
+                    : 'Refresh Active Tokens'
+            }
+            enterDelay={500}
+        >
+            <div className={styles.refresh_container}>
+                <button
+                    className={styles.refresh_button}
+                    onClick={() => handleRefresh()}
+                >
+                    <LuRefreshCcw size={20} />
+                </button>
+            </div>
+        </DefaultTooltip>
+    </div>
+    )
 
-  
+    function handleToggle() {
+        changeView(view);
+        setSearchQueryToken('');
+        setSearchQueryPool('');
+    }
 
     return (
         <section className={styles.main_container}>
-            <div className={styles.main_wrapper}>
-                <h2 className={styles.title_text}>{titleTextForDOM}</h2>
-            </div>
+            {/* <div className={styles.main_wrapper}>
+            <h2 className={styles.title_text}>{titleTextForDOM}</h2>
+            </div> */}
             <div className={styles.options_wrapper}>
-                <div className={styles.options_content}>
-                    <Text>Pools</Text>
-                    <Toggle
-                        isOn={view === 'tokens'}
-                        id={'explore_page_'}
-                        handleToggle={() => {
-                            changeView(view);
-            
-                                setSearchQueryToken('');
-                                setSearchQueryPool('');
-                            
-                        }}
-                    />
-
-                    <Text>Tokens</Text>
-                </div>
-
-                <div className={styles.options_content}>
-                    {inputContainer}
-                    {view === 'pools' && (
-                        <DefaultTooltip
-                            interactive
-                            title={
-                                isExploreDollarizationEnabled
-                                    ? 'Switch to prices in native currency'
-                                    : 'Switch to prices in USD'
-                            }
-                            enterDelay={500}
-                        >
-                            <div className={styles.refresh_container}>
-                                <button
-                                    className={styles.refresh_button}
-                                    onClick={() =>
-                                        setIsExploreDollarizationEnabled(
-                                            (prev) => !prev,
-                                        )
-                                    }
-                                >
-                                    {
-                                        <AiOutlineDollarCircle
-                                            size={20}
-                                            id='trade_dollarized_prices_button'
-                                            aria-label='Toggle dollarized prices button'
-                                            style={{
-                                                color: isExploreDollarizationEnabled
-                                                    ? 'var(--accent1)'
-                                                    : undefined,
-                                            }}
-                                        />
-                                    }
-                                </button>
-                            </div>
-                        </DefaultTooltip>
-                    )}
-                    <DefaultTooltip
-                        interactive
-                        title={
-                            view === 'pools'
-                                ? 'Refresh Top Pools'
-                                : 'Refresh Active Tokens'
-                        }
-                        enterDelay={500}
-                    >
-                        <div className={styles.refresh_container}>
-                            <button
-                                className={styles.refresh_button}
-                                onClick={() => handleRefresh()}
-                            >
-                                <LuRefreshCcw size={20} />
-                            </button>
-                        </div>
-                    </DefaultTooltip>
-                </div>
+           
+            <h2 className={styles.title_text}>{titleTextForDOM}</h2>
+           {optionsContent}
             </div>
 
             {view === 'pools' && (
@@ -283,6 +273,8 @@ export default function Explore(props: ExploreIF) {
                     }
                     searchQuery={searchQueryPool}
                     setSearchQuery={setSearchQueryPool}
+                    view={view}
+                    handleToggle={handleToggle}
                 />
             )}
             {view === 'tokens' && (
@@ -292,6 +284,8 @@ export default function Explore(props: ExploreIF) {
                     goToMarket={goToMarket}
                     searchQuery={searchQueryToken}
                     setSearchQuery={setSearchQueryToken}
+                    view={view}
+                    handleToggle={handleToggle}
                 />
             )}
         </section>
